+++ cloudant-client/src/test/java/com/cloudant/tests/CloudantClientHelper.java
@@ -16,7 +16,7 @@ package com.cloudant.tests;
 
 import com.cloudant.client.api.ClientBuilder;
 import com.cloudant.client.api.CloudantClient;
-import com.squareup.okhttp.mockwebserver.MockWebServer;
+import okhttp3.mockwebserver.MockWebServer;
 
 import java.net.MalformedURLException;
 import java.net.URL;
@@ -112,6 +112,7 @@ public abstract class CloudantClientHelper {
 
     public static ClientBuilder newMockWebServerClientBuilder(MockWebServer mockServer) throws
             MalformedURLException {
+        System.out.println("URL: " + mockServer.url("/"));
         return ClientBuilder.url(mockServer.url("/").url());
     }