+++ cloudant-client/src/test/java/com/cloudant/tests/HttpTest.java
@@ -43,9 +43,6 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonNull;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonPrimitive;
-import com.squareup.okhttp.mockwebserver.MockResponse;
-import com.squareup.okhttp.mockwebserver.MockWebServer;
-import com.squareup.okhttp.mockwebserver.RecordedRequest;
 
 import org.junit.Before;
 import org.junit.ClassRule;
@@ -58,6 +55,10 @@ import org.junit.runners.Parameterized;
 
 import mockit.Mock;
 import mockit.MockUp;
+
+import okhttp3.mockwebserver.MockResponse;
+import okhttp3.mockwebserver.MockWebServer;
+import okhttp3.mockwebserver.RecordedRequest;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;