+++ cloudant-client/src/test/java/com/cloudant/tests/util/MockWebServerResources.java
@@ -14,9 +14,9 @@
 
 package com.cloudant.tests.util;
 
-import com.squareup.okhttp.mockwebserver.Dispatcher;
-import com.squareup.okhttp.mockwebserver.MockResponse;
-import com.squareup.okhttp.mockwebserver.RecordedRequest;
+import okhttp3.mockwebserver.Dispatcher;
+import okhttp3.mockwebserver.MockResponse;
+import okhttp3.mockwebserver.RecordedRequest;
 
 import java.io.FileInputStream;
 import java.security.KeyStore;